requirements.txt
asgiref==3.8.1
boto3==1.36.11
botocore==1.36.11
Django==5.1.2
django-cors-headers==4.5.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
jmespath==1.0.1
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-version==0.0.2
pytz==2024.2
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0